improvement(hdrh): move 'CSHDRFileLogger' to a common place
We are going to add support of hdr histogram files for other
benchmarking tools too.
So, move the 'CSHDRFileLogger' class to a common place
and rename it to the 'HDRHistogramFileLogger'.
